fix: nuqs useSearchParams Suspense boundary 에러 해결
- page.tsx를 server component wrapper로 분리 (Suspense 감싸기)
- 실제 client component를 _components/ 하위로 이동
- clubroom, equipment/[id] 두 페이지 모두 적용
